fix auth integration and fix folder structure
