+++ components/CustomButton/index.jsx
@@ -33,24 +33,26 @@ function LargeButton({ id, label, onClick, loading, className }) {
   );
 }
 
-function GoogleButton({ id, label, onClick, loading }) {
+function GoogleButton({ id, label, onClick, loading, href }) {
   return (
-    <button
-      id={id}
-      onClick={onClick}
-      disabled={loading}
-      className="w-72 md:w-full lg:w-full"
-    >
-      <div className=" bg-blue-500 flex text-white border-2 border-blue-500 items-center font-Poppins text-lg md:text-2xl lg:text-2xl font-medium">
-        <div className="bg-white p-1 flex md:hidden lg:hidden">
-          <Image src="/logo-google.png" alt="logo" width={20} height={20} />
+    <Link href={`${href}`}>
+      <button
+        id={id}
+        onClick={onClick}
+        disabled={loading}
+        className="w-72 md:w-full lg:w-full"
+      >
+        <div className=" bg-blue-500 flex text-white border-2 border-blue-500 items-center font-Poppins text-lg md:text-2xl lg:text-2xl font-medium">
+          <div className="bg-white p-1 flex md:hidden lg:hidden">
+            <Image src="/logo-google.png" alt="logo" width={20} height={20} />
+          </div>
+          <div className="bg-white p-1 hidden md:flex lg:flex">
+            <Image src="/logo-google.png" alt="logo" width={40} height={40} />
+          </div>
+          <div className="w-full px-3 flex justify-center">{label}</div>
         </div>
-        <div className="bg-white p-1 hidden md:flex lg:flex">
-          <Image src="/logo-google.png" alt="logo" width={40} height={40} />
-        </div>
-        <div className="w-full px-3 flex justify-center">{label}</div>
-      </div>
-    </button>
+      </button>
+    </Link>
   );
 }
