+++ styles/globals.css
@@ -15,3 +15,11 @@ body {
 * {
   box-sizing: border-box;
 }
+
+input[type="date"]::-webkit-calendar-picker-indicator {
+  /* background-color: #ffc700; */
+  background-color: white;
+}
+input[type="time"]::-webkit-calendar-picker-indicator {
+  background-color: white;
+}
